modularise new home page
